Separated main into modules
